refactor: move decode lib logs to ecosystem
