+++ src/ape_ethereum/transactions.py
@@ -8,14 +8,13 @@ from eth_account._utils.legacy_transactions import (
     encode_transaction,
     serializable_unsigned_transaction_from_dict,
 )
-from eth_utils import decode_hex, keccak, to_int
+from eth_utils import keccak, to_int
 from ethpm_types import HexBytes
 from pydantic import BaseModel, Field, root_validator, validator
 from rich.console import Console as RichConsole
 
 from ape.api import ReceiptAPI, TransactionAPI
 from ape.exceptions import OutOfGasError, SignatureError, TransactionError
-from ape.types import ContractLog
 from ape.utils import CallTraceParser, TraceStyles
 
 
@@ -156,47 +155,6 @@ class Receipt(ReceiptAPI):
         elif self.status != TransactionStatusEnum.NO_ERROR:
             txn_hash = HexBytes(self.txn_hash).hex()
             raise TransactionError(message=f"Transaction '{txn_hash}' failed.")
-
-    def decode_library_log(self, log: Dict) -> Optional[ContractLog]:
-        return self._decode_ds_note(log)
-
-    def _decode_ds_note(self, log: Dict) -> Optional[ContractLog]:
-        """
-        Decode anonymous events emitted by the DSNote library.
-        """
-
-        # The first topic encodes the function selector
-        selector, tail = log["topics"][0][:4], log["topics"][0][4:]
-        if sum(tail):
-            # non-zero bytes found after selector
-            return None
-
-        contract_type = self.chain_manager.contracts.get(log["address"])
-        if contract_type is None:
-            # contract type for {log['address']} not found
-            return None
-
-        try:
-            method_abi = contract_type.mutable_methods[selector]
-        except KeyError:
-            #  selector {selector.hex()} not found in {log['address']}
-            return None
-
-        # ds-note data field uses either (uint256,bytes) or (bytes) encoding
-        # instead of guessing, assume the payload begins right after the selector
-        data = decode_hex(log["data"])
-        input_types = [i.canonical_type for i in method_abi.inputs]
-        start_index = data.index(selector) + 4
-        values = decode_abi(input_types, data[start_index:])
-
-        return ContractLog(  # type: ignore
-            name=method_abi.name,
-            block_hash=log["blockHash"],
-            block_number=log["blockNumber"],
-            event_arguments={i.name: value for i, value in zip(method_abi.inputs, values)},
-            index=log["logIndex"],
-            transaction_hash=log["transactionHash"],
-        )
 
     def show_trace(self, verbose: bool = False, file: IO[str] = sys.stdout):
         tree_factory = CallTraceParser(self, verbose=verbose)
