+++ tests/functional/conftest.py
@@ -275,6 +275,7 @@ def ds_note():
     }
 
 
+@pytest.fixture
 def remove_disk_writes_deployments(chain):
     if chain.contracts._deployments_mapping_cache.exists():
         chain.contracts._deployments_mapping_cache.unlink()
